// app/generated-api/models/solution.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Solutions API
 * REST API for technical solutions organized by categories and subcategories
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Solution
 */
export interface Solution {
    /**
     * 
     * @type {any}
     * @memberof Solution
     */
    'id'?: any;
    /**
     * 
     * @type {any}
     * @memberof Solution
     */
    'title'?: any;
    /**
     * 
     * @type {any}
     * @memberof Solution
     */
    'content'?: any;
    /**
     * 
     * @type {any}
     * @memberof Solution
     */
    'difficulty'?: SolutionDifficultyEnum;
    /**
     * 
     * @type {any}
     * @memberof Solution
     */
    'helpful_count'?: any;
    /**
     * 
     * @type {any}
     * @memberof Solution
     */
    'subcategory_id'?: any;
    /**
     * 
     * @type {any}
     * @memberof Solution
     */
    'created_at'?: any;
}

export const SolutionDifficultyEnum = {
    Beginner: 'beginner',
    Intermediate: 'intermediate',
    Advanced: 'advanced'
} as const;

export type SolutionDifficultyEnum = typeof SolutionDifficultyEnum[keyof typeof SolutionDifficultyEnum];